Initial commit
updated some of the looks/colors in the stylesheets.

rearranged some functions and styles to go out of their individual
modules and into the STcommon files instead

Started documenting some functions

There now is a concept of showLen(gth).  We're still only tracking one
event for now, the showBegin, but we figure if the current time is
after the show and increment the day by one if show's over (assuminmg
the show begins tomorrow instead of today).