Add area to bottom of table for messages
This area is intended to be used for reminders, such as billboards or
live reads.  It's unpopulated now except for a nonbreaking space, but
eventually it's probably going to have a dimissal button.  It may even
gain its own .st structure so it can be blinked.  Also includes some
styling in the .css file.