Added some code for automatic timezone selection
The original was pegged to US/Eastern time (show begins at 14:06).
Now the getTimezoneOffset() method of the Date object helps figure the
local timezone relative to the Premiere chart, which is primarily in
US/Eastern (but mentions PT).  The show of course originates from
US/Pacific, so starts at 11:06.  The program should now work at least
in the main 48 states, not sure, perhaps AK and HI too.  Not
guaranteed to work anywhere else because daylight saving time rules
might be different from that of "The Tech Guy" show.