initial show status bugfix; turn off blinking
The loop that tries to determine where the present time fits into the breaks list did not check for SHOW_DONE, resulting in behavior where if the showRef (usuall OTA) time was close to the end, JS would throw an error about indexing off the end of etab[]

The SHOW_DONE state generally leaves the verysoon setInterval() going, thus blinking (by default) red.  Now we go through every time object (tmobj) and every color descriptor and call endBlink() when stopping.  This should be safe as the .timer member is initialized to null in the constructor and endBlink checks for its argument being null and doesn't try to clearInterval() if so.