Added code to strip "var defaultbrk = " from JSON
The idea is to be able to add a <script> tag to a page and define the "default" break table, but still be able to have people copy/paste from the example .json file into the import/export area.  The techguy.json file is thus altered so that can be so.