refactor load_breaks() to have a default array
The HTML file now has another <script> tag which includes
"techguy.json", which is expected to define "defaultbrk". If the
requested profile (by default "techguy_breaks") can't be gotten out of
localStorage (for example on Google Chrome which does not allow
file:/// URIs to use localStorage.getItem()), if "defaultbrk" is an
object (and arrays are objects), return that instead.  So in effect
that provides a builtin set of break descriptors.