centralized more fns into STcommon.js, color revertCSS added
find_all_profiles() and get_show_start_hour_offset() moved into
STcommon because they'll ultimately be useful to all modules.

Added a member to the st.color member with all its colors set to the
null string.  If the color or background is the null string, those
style aspects revert to CSS.  That simplified choosing a color member
based on the show state.  Instead of having to test separately for
null string return, "revertCSS" is returned instead, so a straight
color[st2col()] can be referenced/called.