arg list change for dbg()
Some browsers, for example Dolphin/Android, do not understand ellipsis
notation for variable number of arguments, so this change was backed
out.  The whole reason for trying to use it is the overloading of "+"
as both an addition operator and a string concatenator.  A string
added to a number, then some other numeric operation (say,
subtraction) would result in NaN instead of string concat.