minor code refactor; expose global timing; OOS for countdown
Break out a few places which were changing colors based on state or
new state, this is now in a separate function.

The main countdown timer now also goes (by default) pink for an out of
sync event.

The globalRealOff is exposed to an input blank now, so that instant
timing updates can be made w/o having to pull up the dev console and
setting it (handy for syncing audio-only stream to our clock, which is
about a factor of -10000).

The default background for really soon blink is maroon instead of red
(toned it down a notch visually)