Timer page elements generalized
Timer objects are now discovered with querySelectorAll() for
data-st-type=timer.  As the comment notes, in JavaScript, properties
of window become unqualified variable names.  So for example
window["tmtilbreak"] becomes the variable tmtilbreak as if on the
topmost scope one were to write "var tmtilbreak;".  So the HTML had to
be adjusted accordingly; some of the IDs were descriptive mixed case
and longer names, but now the id attribute matches the variable name.

Some bizzarre errors were popping up in the Firefox console about
violations of content security policy, so a meta http-equiv element
was added.  Don't know where that came from, because things were
operating fine the day before this commit was made, and nothing which
would seem to affect JavaScript interpretation was changed.

The default timing tolerance was increased from 75 to 100 ms.

Some debugging statement levels were changed now that that section has
been debugged.

The slow tick does not update nxtbreaktm each second anymore, only on
state transitions and in ST_init()

ST_init() now clears out the event table etab instead of it being done
on the outermost level.  That means it's easier to set showBegin then
call ST_init() in the developer's console.