Fixed bug with stop function crashing
The tmobj array was getting a NodeList, and that list has not only
references to nodes in an array, but also methods and properties.  The
shutdown routine was choking on "item()" when endBlink()ing (as item()
does not have our "st" structure).  So in ST_init(), copy the list of
nodes to tmobj instead of just assigning it from "timernode"