Added files for reminder support
The editing code, both HTML and JavaScript, was mostly copied stuff
from ShowTimerBreakEdit.  Names were changed to be sane, and instead
of all the miscellany needed for end times (calculations, checking for
overlap, etc.) became quite unnecessary, and instead it caters to just
pairing a text message with the begin time.