Reformatted README.md to wrap long lines
I found out that as long as newlines aren't doubled (a blank line),
Markdown does not care.  I thought because GitHub's editor did not
insert line breaks that this would cause the same line wrapping in a
browser.  This is not so, text is rewrapped to suit whatever
dimensions the browser wants to accommodate; there are no extra line
break tags inserted just because text is wrapped at column 70 or
whatever.  Editing with Emacs or vi is much easier when wrapped;
otherwise editing within a paragraph is a lot of skipping back or
forth by words.