small improvements to ShowTimer.js
The st object now gets all elements with the attribute "data-st-role"
added to it in the constructor function, instead of a bunch of
explicit this.<something> = HTelement (basically uses the construct
this[getAttribute()] to instantiate it).

There is now the function start_soon() which takes and argument
(default 2) of how many minutes into the future to set the show
beginning.  Eventually this might be attached to a button or other
input field so that the host can do e.g. prerecords and use the same
timing.  For now, it facilitates testing of the show open by calling
it in the debugger/devtools.