no active JavaScript warning added
The author has been writing and saying for years that putting
prominent messages on Web pages informing the user that JavaScript is
not active is so trivial, every Web page where it's critical to its
function should have one.  This set of changes removes the author from
the hypocrite category.