preparations for generalizing any show
Added some title texts for the main page so that on pointer hover the
purpose of the region will be shown to the user.

Catch any JSON error when importing breaks.  Also, the prompt for
overwriting existing breaks on the page is deferred until JSON.parse()
succeeds.

Any place a <datalist> is populated, what is added is filtered by the
value of the data-st-proftype attribute, so that only breaks are
included where breaks are being sought, and so forth for params and
reminders.  The search routine (find_all_profiles()) still won't
filter on _bak, _bak0, _bak1, etc., that is up to the caller to
decide.  (Besides, at the time of this commit, multiple backups are
not implemented, only one.  That'd only be for experts working from
the debugging console.)