Removed some constructs not known to some browsers
While running this on Dolphin Browser, it complained (syntax errors)
about using the for ( v of arr ) construct.  There is a loop iterating
through the same array a few lines later anyway, so the fix for that
was easy, just move the code in that block into the other block.

It also did not understand the syntax fn(v = d) (default for function
formal parameters), so I removed that construct from the parameter
list of start_soon() (default to start in two minutes).